+++ qupath-gui-fx/src/main/java/qupath/lib/gui/commands/BrightnessContrastCommand.java
@@ -23,12 +23,14 @@
 
 package qupath.lib.gui.commands;
 
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonSyntaxException;
 import javafx.application.Platform;
 import javafx.beans.InvalidationListener;
 import javafx.beans.Observable;
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.BooleanBinding;
-import javafx.beans.binding.ListBinding;
 import javafx.beans.binding.ObjectExpression;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
@@ -51,7 +53,6 @@ import javafx.scene.control.Tooltip;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.Priority;
@@ -64,6 +65,7 @@ import org.slf4j.LoggerFactory;
 import qupath.fx.utils.GridPaneUtils;
 import qupath.lib.display.ChannelDisplayInfo;
 import qupath.lib.display.ImageDisplay;
+import qupath.lib.display.settings.DisplaySettingUtils;
 import qupath.lib.display.settings.ImageDisplaySettings;
 import qupath.lib.gui.QuPathGUI;
 import qupath.lib.gui.actions.InfoMessage;
@@ -74,10 +76,14 @@ import qupath.lib.gui.commands.display.BrightnessContrastSliderPane;
 import qupath.lib.gui.prefs.PathPrefs;
 import qupath.lib.gui.viewer.QuPathViewer;
 import qupath.lib.images.ImageData;
+import qupath.lib.io.GsonTools;
 
 import java.awt.image.BufferedImage;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
+import java.io.File;
+import java.nio.file.Files;
+import java.util.List;
 import java.util.Objects;
 
 /**
@@ -144,6 +150,33 @@ public class BrightnessContrastCommand implements Runnable {
 		chartPane.updateHistogram(imageDisplayProperty.get(), currentChannelProperty.get());
 
 		dialog = createDialog();
+
+		qupath.getDefaultDragDropListener().addJsonDropHandler(this::handleDisplaySettingsDrop);
+	}
+
+	private boolean handleDisplaySettingsDrop(QuPathViewer viewer, List<JsonElement> elements) {
+		if (elements.size() != 1 || viewer == null || viewer.getImageDisplay() == null)
+			return false;
+		var element = elements.get(0);
+		if (element.isJsonObject()) {
+			var obj = element.getAsJsonObject();
+			if (obj.has("name") && obj.has("channels") && obj.has("gamma")) {
+				var gson = GsonTools.getInstance();
+				try {
+					var settings = gson.fromJson(element, ImageDisplaySettings.class);
+					if (DisplaySettingUtils.applySettingsToDisplay(viewer.getImageDisplay(), settings)) {
+						maybeSyncSettingsAcrossViewers(viewer.getImageDisplay());
+						logger.info("Applied display settings: {}", settings.getName());
+					} else {
+						logger.warn("Unable to apply display settings: {}", settings.getName());
+					}
+					return true;
+				} catch (JsonSyntaxException e) {
+					logger.debug("Unable to parse display settings", e);
+				}
+			}
+		}
+		return false;
 	}
